Fix shaded dependencies
Remove Kotlin stdlib from shaded libraries, which were not relocated
causing it to conflict with other dependencies. Setting
promoteTransitiveDependencies ensure the Kotlin stdlib, being a
dependency of okhttp3, will be listed as a dependency and installed
by consumers.

Create and use a reduced pom, so the shaded dependencies are not also
specified as dependencies.