Support trailing slash in base URL.
While there is technically a difference between a URL that ends in a slash and one that does not, it is so common to supply a trailing slash in base URLs that this is a substantial trap, which is hard to debug because the resulting authentication error is quite opaque (I just wasted two hours because of this). Furthermore, the Python SDK already accepts a trailing slash.